requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
distlib==0.3.8
Django==5.0.7
django-auditlog==3.0.0
django-core==1.4.1
django-cors-headers==4.4.0
django-filter==24.2
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
et-xmlfile==1.1.0
filelock==3.15.1
gunicorn==22.0.0
idna==3.7
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.4.0
platformdirs==4.2.2
psycopg==3.2.1
psycopg2-binary==2.9.9
pycparser==2.22
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.2
